requirements.txt
asgiref==3.2.7
autopep8==1.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-csvimport==2.15
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-solo==1.1.3
gunicorn==20.0.4
psycopg2==2.8.5
pycodestyle==2.5.0
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
